# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  AccountActivationRequestBody:
    properties:
      activation_code:
        description: activation code obtained from registration process (sent via
          email)
        type: string
        x-go-name: ActivationCode
    required:
    - activation_code
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/dto
...
  DefaultResponseBody:
    properties:
      message:
        description: The response message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/dto
...
  SignInResponseBody:
    properties:
      message:
        description: The response message
        example: signin succeeded
        type: string
        x-go-name: Message
      token:
        $ref: '#/definitions/Token'
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/dto
  Token:
    properties:
      access_token:
        description: access token
        type: string
        x-go-name: AccessToken
      refresh_token:
        description: refresh token
        type: string
        x-go-name: RefreshToken
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/entities
  ValidationError:
    properties:
      field:
        type: string
        x-go-name: Field
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/lib
  ValidationErrorResponseBody:
    properties:
      errors:
        description: Field errors with its messages
        items:
          $ref: '#/definitions/ValidationError'
        type: array
        x-go-name: Errors
      message:
        description: The message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/ilhamtubagus/urlShortener/dto
host: localhost
info:
  contact:
    email: ilhamta27@gmail.com
    name: Ilham Tubagus Arfian
    url: https://github.com/ilhamtubagus
  description: |-
    Documentation for Shorkeun API.
    Why do we need to shorten the URL? Is it something necessary?
    Well, there are many advantages that shortening of URL provides.
    A very basic advantage could be that users tend to make very few mistakes while copying the URL if it is not shortened.
    Secondly, they surely save a lot of space when used or printed.
    Moreover, it could be used if someone wishes not to use the original URL or want to hide the original one.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: There are no TOS at this moment, use at your own risk we take no
    responsibility
  title: Shorkeun API.
  version: 1.0.0
paths:
...
  /user/status:
    patch:
      consumes:
      - application/json
      description: Activate user's account with activation code sent via email. User's
        status will change to "ACTIVE"
      operationId: accountActivation
      parameters:
      - in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/AccountActivationRequestBody'
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/signInResponse'
        "400":
          $ref: '#/responses/defaultResponse'
        "404":
          $ref: '#/responses/defaultResponse'
        "422":
          $ref: '#/responses/validationError'
        "500":
          $ref: '#/responses/defaultResponse'
      security:
      - Bearer-Token: []
      summary: Account activation
      tags:
      - user
produces:
- application/json
responses:
  defaultResponse:
    description: A default response with message that describe the response result
    schema:
      $ref: '#/definitions/DefaultResponseBody'
  signInResponse:
    description: A response when user's successfully signed in
    schema:
      $ref: '#/definitions/SignInResponseBody'
  validationError:
    description: A ValidationError is an error that is used when the required input
      fails validation.
    schema:
      $ref: '#/definitions/ValidationErrorResponseBody'
schemes:
- http
- https
securityDefinitions:
  Bearer-Token:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"